requirements-dev.txt
boto3==1.9.91
botocore==1.12.91
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
django-filter==2.1.0
django-storages==1.7.1
djangorestframework==3.9.1
docutils==0.14
jmespath==0.9.3
Markdown==3.0.1
Pillow==5.4.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
static3==0.7.0
urllib3==1.24.1
